+++ app/components/ImportOPMLDialog.vue
@@ -65,7 +65,7 @@ const file = ref<File | null>(null)
 const onFileChange = (event: Event) => {
   const target = event.target as HTMLInputElement
   if (target.files && target.files.length > 0) {
-    file.value = target.files[0]
+    file.value = target.files[0] ?? null
   }
 }
