+++ server/utils/podParser/types/index.ts
@@ -1,3 +1,1 @@
 export * from './ParsedPodcastXML'
-export * from './Podcast'
-export * from './Episode'
