--- shared/types/Edpisode.ts
@@ -1,1 +1,0 @@
-export type { Episode } from '../server/utils/podParser/types/Episode'
