+++ shared/types/Podcast.ts
@@ -1,1 +1,30 @@
-export type { Podcast } from '../server/utils/podParser/types/Podcast'
+import type { PodcastFunding } from './PodcastFunding'
+import type { PodcastPerson } from './PodcastPerson'
+
+export interface Podcast {
+  copyright: string
+  contentEncoded: string
+  description: string
+  feedUrl: string
+  image: {
+    link: string
+    title: string
+    url: string
+  }
+  itunesAuthor: string
+  itunesCategory: string
+  itunesExplicit: boolean
+  itunesImage: string
+  itunesOwner: {
+    name: string
+    email: string
+  }
+  itunesSubtitle: string
+  itunesSummary: string
+  itunesType: string
+  language: string
+  link: string
+  title: string
+  persons?: PodcastPerson[]
+  funding?: PodcastFunding[]
+}
